# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application API for project WineApp",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/vineyard/{vineyardId}": {
            "get": {
                "tags": [
                    "Vineyard"
                ],
                "summary": "Show vineyard by vineyardId",
                "description": "Show vineyard by vineyardId",
                "operationId": "showVineyardById",
                "parameters": [
                    {
                        "name": "vineyardId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Vineyard"
                ],
                "summary": "Update Vineyard",
                "description": "Update Vineyard by selected vineyardId",
                "operationId": "updateVineyard",
                "parameters": [
                    {
                        "name": "vineyardId",
                        "in": "path",
                        "description": "Vineyard ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string"
                                    },
                                    "area": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/vineyard/project/{projectId}": {
            "get": {
                "tags": [
                    "Vineyard"
                ],
                "summary": "Show vineyards by projectId",
                "description": "Show vineyards by projectId",
                "operationId": "showVineyardsByProjectId",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/vineyard": {
            "post": {
                "tags": [
                    "Vineyard"
                ],
                "summary": "Create Vineyard",
                "description": "Create Vineyard in project",
                "operationId": "createVineyard",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "title"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    },
                                    "area": {
                                        "type": "number",
                                        "format": "double"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}